web-animations/animation-timeline/animation-timeline-current-time.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationTimeline currentTime attribute tests</title>
<meta name="assert" content="AnimationTimeline currentTime attribute returns the time value for this timeline">
<link rel="help" href="http://dev.w3.org/fxtf/web-animations/#the-animationtimeline-interface">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
test(function() {
    assert_not_equals(document.timeline.currentTime, null,
        'Time value for the started AnimationTimeline should not be null');
}, 'Test started AnimationTimeline currentTime is not null');

test(function() {
    assert_greater_than(document.timeline.currentTime, 0,
        'Time value for the timeline should be greater than zero');
}, 'Timeline value is greater than zero test');
</script>
</body>
